remove cal

+++ src/QmlRobotData.h
@@ -19,9 +19,9 @@ class QmlMcuData : public QObject
   Q_PROPERTY(QVector<int> motorsCcr MEMBER motorsCcr NOTIFY updated)
   Q_PROPERTY(QVector<float> battery1 MEMBER battery1 NOTIFY updated)
   Q_PROPERTY(QVector<float> battery2 MEMBER battery2 NOTIFY updated)
-  Q_PROPERTY(QVector<double> accelerometer MEMBER accelerometer NOTIFY updated)
-  Q_PROPERTY(QVector<double> gyroscope MEMBER gyroscope NOTIFY updated)
-  Q_PROPERTY(QVector<double> magnetometer MEMBER magnetometer NOTIFY updated)
+  Q_PROPERTY(QVector<float> accelerometer MEMBER accelerometer NOTIFY updated)
+  Q_PROPERTY(QVector<float> gyroscope MEMBER gyroscope NOTIFY updated)
+  Q_PROPERTY(QVector<float> magnetometer MEMBER magnetometer NOTIFY updated)
   Q_PROPERTY(QVector<float> hardIronMax MEMBER hardIronMax NOTIFY updated)
   Q_PROPERTY(QVector<float> hardIronMin MEMBER hardIronMin NOTIFY updated)
   Q_PROPERTY(QVector<float> softIronMatrix MEMBER softIronMatrix NOTIFY updated)
@@ -40,9 +40,9 @@ class QmlMcuData : public QObject
     QVector<int> motorsCcr = {0, 0, 0, 0}; // motor 1, motor 2, motor 3, motor 4
     QVector<float> battery1 = {0.0f, 0.0f}; // voltage, current
     QVector<float> battery2 = {0.0f, 0.0f}; // voltage, current
-    QVector<double> accelerometer = {0.0f, 0.0f, 0.0f}; // x, y, z
-    QVector<double> gyroscope = {0.0f, 0.0f, 0.0f}; // x, y, z
-    QVector<double> magnetometer = {0.0f, 0.0f, 0.0f}; // x, y, z
+    QVector<float> accelerometer = {0.0f, 0.0f, 0.0f}; // x, y, z
+    QVector<float> gyroscope = {0.0f, 0.0f, 0.0f}; // x, y, z
+    QVector<float> magnetometer = {0.0f, 0.0f, 0.0f}; // x, y, z
     QVector<float> hardIronMax = {0.0f, 0.0f, 0.0f}; // x, y, z
     QVector<float> hardIronMin = {0.0f, 0.0f, 0.0f}; // x, y, z
     QVector<float> softIronMatrix = {0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f, 0.0f}; // Matrix
